requirements.txt
streamlit
pandas
openpyxl
xlrd
python-dotenv
mysql-connector-python
